pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beyondml"
version = "0.1.0"
description = "AI Agent Orchestration Platform — Terminal-native AutoML with LLM-powered agents and Genetic Algorithm optimization."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Riteesh", email = "riteesh@example.com"},
]
keywords = ["automl", "machine-learning", "genetic-algorithm", "llm", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "textual>=0.40",
    "plotext>=5.2",
    "python-dotenv>=1.0",
    "groq>=0.4",
    "requests>=2.31",
    "pmlb>=1.0",
    "joblib>=1.3",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "coverage>=7.0",
]

[project.scripts]
beyondml = "beyondml.cli:main"

[tool.setuptools.packages.find]
include = ["beyondml*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
